(merged from branches/roa. use "svn log -c <changeset> -g <module-svn-path>" for detailed commit message)
git-svn-id: svn://svn.silverstripe.com/silverstripe/open/modules/sapphire/trunk@60235 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ api/RestfulServer.php
@@ -12,10 +12,10 @@
  * applications.
  * 
  *  - GET /api/v1/(ClassName)/(ID) - gets a database record
- *  - GET /api/v1/(ClassName)/(ID)/(Relation) - get all of the records linked to this database record by the given reatlion (NOT IMPLEMENTED YET)
- *  - GET /api/v1/(ClassName)?(Field)=(Val)&(Field)=(Val) - searches for matching database records (NOT IMPLEMENTED YET)
+ *  - GET /api/v1/(ClassName)/(ID)/(Relation) - get all of the records linked to this database record by the given reatlion
+ *  - GET /api/v1/(ClassName)?(Field)=(Val)&(Field)=(Val) - searches for matching database records
  * 
- *  - PUT /api/v1/(ClassName)/(ID) - updates a database record (NOT IMPLEMENTED YET)
+ *  - PUT /api/v1/(ClassName)/(ID) - updates a database record
  *  - PUT /api/v1/(ClassName)/(ID)/(Relation) - updates a relation, replacing the existing record(s) (NOT IMPLEMENTED YET)
  *  - POST /api/v1/(ClassName)/(ID)/(Relation) - updates a relation, appending to the existing record(s) (NOT IMPLEMENTED YET)
  * 
@@ -34,8 +34,10 @@
  * - &fields=<string>: Comma-separated list of fields on the output object (defaults to all database-columns)
  * 
  * @todo Finish RestfulServer_Item and RestfulServer_List implementation and re-enable $url_handlers
+ * @todo Implement PUT/POST/DELETE for relations
  * @todo Make SearchContext specification customizeable for each class
  * @todo Allow for range-searches (e.g. on Created column)
+ * @todo Allow other authentication methods (currently only HTTP BasicAuth)
  */
 class RestfulServer extends Controller {
 	static $url_handlers = array(
@@ -52,7 +54,7 @@ class RestfulServer extends Controller {
 	 *
 	 * @var string
 	 */
-	protected static $default_extension = "xml";
+	public static $default_extension = "xml";
 	
 	/**
 	 * If no extension is given, resolve the request to this mimetype.
@@ -60,19 +62,6 @@ class RestfulServer extends Controller {
 	 * @var string
 	 */
 	protected static $default_mimetype = "text/xml";
-	
-	/**
-	 * Maps common extensions to their mimetype representations.
-	 *
-	 * @var array
-	 */
-	protected static $mimetype_map = array(
-		'xml' => 'text/xml',
-		'json' => 'text/json',
-		'js' => 'text/json',
-		'xhtml' => 'text/html',
-		'html' => 'text/html',
-	);
 	
 	/*
 	function handleItem($request) {
@@ -91,24 +80,25 @@ class RestfulServer extends Controller {
 	function index() {
 		ContentNegotiator::disable();
 		
-		$requestMethod = $_SERVER['REQUEST_METHOD'];
 		if(!isset($this->urlParams['ClassName'])) return $this->notFound();
 		$className = $this->urlParams['ClassName'];
 		$id = (isset($this->urlParams['ID'])) ? $this->urlParams['ID'] : null;
 		$relation = (isset($this->urlParams['Relation'])) ? $this->urlParams['Relation'] : null;
 		
-		switch($requestMethod) {
-			case 'GET':
-				return $this->getHandler($className, $id, $relation);
-			
-			case 'PUT':
-				return $this->putHandler($className, $id, $relation);
-			
-			case 'DELETE':
-				return $this->deleteHandler($className, $id, $relation);
-			
-			case 'POST':
-		}
+		// if api access is disabled, don't proceed
+		if(!singleton($className)->stat('api_access')) return $this->permissionFailure();
+		
+		// authenticate through HTTP BasicAuth
+		$member = $this->authenticate();
+
+		// handle different HTTP verbs
+		if($this->request->isGET()) return $this->getHandler($className, $id, $relation);
+		if($this->request->isPOST()) return $this->postHandler($className, $id, $relation);
+		if($this->request->isPUT()) return $this->putHandler($className, $id, $relation);
+		if($this->request->isDELETE()) return $this->deleteHandler($className, $id, $relation);
+
+		// if no HTTP verb matches, return error
+		return $this->methodNotAllowed();
 	}
 	
 	/**
@@ -137,6 +127,8 @@ class RestfulServer extends Controller {
 	 *   - static $api_access must be set. This enables the API on a class by class basis
 	 *   - $obj->canView() must return true. This lets you implement record-level security
 	 * 
+	 * @todo Access checking
+	 * 
 	 * @param String $className
 	 * @param Int $id
 	 * @param String $relation
@@ -152,17 +144,13 @@ class RestfulServer extends Controller {
 			'limit' => $this->request->getVar('limit')
 		);
 		
-		$formatter = $this->getDataFormatter();
-
+		$responseFormatter = $this->getResponseDataFormatter();
+		if(!$responseFormatter) return $this->unsupportedMediaType();
+		
 		if($id) {
 			$obj = DataObject::get_by_id($className, $id);
-			if(!$obj) {
-				return $this->notFound();
-			}
-			
-			if(!$obj->stat('api_access') || !$obj->canView()) {
-				return $this->permissionFailure();
-			}
+			if(!$obj) return $this->notFound();
+			if(!$obj->canView()) return $this->permissionFailure();
 			
 			if($relation) {
 				if($relationClass = $obj->many_many($relation)) {
@@ -185,16 +173,15 @@ class RestfulServer extends Controller {
 			} 
 			
 		} else {
-			if(!singleton($className)->stat('api_access')) {
-				return $this->permissionFailure();
-			}
 			$obj = $this->search($className, $this->request->getVars(), $sort, $limit);
 			// show empty serialized result when no records are present
 			if(!$obj) $obj = new DataObjectSet();
 		}
 		
-		if($obj instanceof DataObjectSet) return $formatter->convertDataObjectSet($obj);
-		else return $formatter->convertDataObject($obj);
+		$this->getResponse()->addHeader('Content-Type', $responseFormatter->getOutputContentType());
+		
+		if($obj instanceof DataObjectSet) return $responseFormatter->convertDataObjectSet($obj);
+		else return $responseFormatter->convertDataObject($obj);
 	}
 	
 	/**
@@ -220,50 +207,147 @@ class RestfulServer extends Controller {
 		return singleton($className)->buildDataObjectSet($query->execute());
 	}
 
-	protected function getDataFormatter() {
+	/**
+	 * Returns a dataformatter instance based on the request
+	 * extension or mimetype. Falls back to {@link self::$default_extension}.
+	 * 
+	 * @param boolean $includeAcceptHeader Determines wether to inspect and prioritize any HTTP Accept headers 
+	 * @return DataFormatter
+	 */
+	protected function getDataFormatter($includeAcceptHeader = false) {
 		$extension = $this->request->getExtension();
+		$contentType = $this->request->getHeader('Content-Type');
+		$accept = $this->request->getHeader('Accept');
+
+		// get formatter
+		if(!empty($extension)) {
+			$formatter = DataFormatter::for_extension($extension);
+		}elseif($includeAcceptHeader && !empty($accept) && $accept != '*/*') {
+			$formatter = DataFormatter::for_mimetypes(explode(',',$accept));
+		} elseif(!empty($contentType)) {
+			$formatter = DataFormatter::for_mimetype($contentType);
+		} else {
+			$formatter = DataFormatter::for_extension(self::$default_extension);
+		}
 		
-		// Determine mime-type from extension
-		$contentType = isset(self::$mimetype_map[$extension]) ? self::$mimetype_map[$extension] : self::$default_mimetype;
-		
-		if(!$extension) $extension = self::$default_extension;
-		$formatter = DataFormatter::for_extension($extension); //$this->dataFormatterFromMime($contentType);
+		// set custom fields
 		if($customFields = $this->request->getVar('fields')) $formatter->setCustomFields(explode(',',$customFields));
+
+		// set relation depth
 		$relationDepth = $this->request->getVar('relationdepth');
 		if(is_numeric($relationDepth)) $formatter->relationDepth = (int)$relationDepth;
 		
 		return $formatter;		
 	}
 	
+	protected function getRequestDataFormatter() {
+		return $this->getDataFormatter(false);
+	}
+	
+	protected function getResponseDataFormatter() {
+		return $this->getDataFormatter(true);
+	}
+	
 	/**
 	 * Handler for object delete
 	 */
 	protected function deleteHandler($className, $id) {
-		if($id) {
-			$obj = DataObject::get_by_id($className, $id);
-			if($obj->stat('api_access') && $obj->canDelete()) {
-				$obj->delete();
-			} else {
-				return $this->permissionFailure();
-			}
-		}
+		$obj = DataObject::get_by_id($className, $id);
+		if(!$obj) return $this->notFound();
+		if(!$obj->canDelete()) return $this->permissionFailure();
+		
+		$obj->delete();
+		
+		$this->getResponse()->setStatusCode(204); // No Content
+		return true;
 	}
 
 	/**
 	 * Handler for object write
 	 */
 	protected function putHandler($className, $id) {
-		return $this->permissionFailure();
+		$obj = DataObject::get_by_id($className, $id);
+		if(!$obj) return $this->notFound();
+		if(!$obj->canEdit()) return $this->permissionFailure();
+		
+		$reqFormatter = $this->getRequestDataFormatter();
+		if(!$reqFormatter) return $this->unsupportedMediaType();
+		
+		$responseFormatter = $this->getResponseDataFormatter();
+		if(!$responseFormatter) return $this->unsupportedMediaType();
+		
+		$obj = $this->updateDataObject($obj, $reqFormatter);
+		
+		$this->getResponse()->setStatusCode(200); // Success
+		$this->getResponse()->addHeader('Content-Type', $responseFormatter->getOutputContentType());
+		$objHref = Director::absoluteURL(self::$api_base . "$obj->class/$obj->ID");
+		$this->getResponse()->addHeader('Location', $objHref);
+		
+		return $responseFormatter->convertDataObject($obj);
 	}
 
 	/**
-	 * Handler for object append / method call
+	 * Handler for object append / method call.
+	 * 
+	 * @todo Posting to an existing URL (without a relation)
+	 * current resolves in creatig a new element,
+	 * rather than a "Conflict" message.
 	 */
 	protected function postHandler($className, $id) {
-		return $this->permissionFailure();
+		if(!singleton($className)->canCreate()) return $this->permissionFailure();
+		$obj = new $className();
+		
+		$reqFormatter = $this->getRequestDataFormatter();
+		if(!$reqFormatter) return $this->unsupportedMediaType();
+		
+		$responseFormatter = $this->getResponseDataFormatter();
+		
+		$obj = $this->updateDataObject($obj, $reqFormatter);
+		
+		$this->getResponse()->setStatusCode(201); // Created
+		$this->getResponse()->addHeader('Content-Type', $responseFormatter->getOutputContentType());
+		$objHref = Director::absoluteURL(self::$api_base . "$obj->class/$obj->ID");
+		$this->getResponse()->addHeader('Location', $objHref);
+		
+		return $responseFormatter->convertDataObject($obj);
+		
 	}
-
-
+	
+	/**
+	 * Converts either the given HTTP Body into an array
+	 * (based on the DataFormatter instance), or returns
+	 * the POST variables.
+	 * Automatically filters out certain critical fields
+	 * that shouldn't be set by the client (e.g. ID).
+	 *
+	 * @param DataObject $obj
+	 * @param DataFormatter $formatter
+	 * @return DataObject The passed object
+	 */
+	protected function updateDataObject($obj, $formatter) {
+		// if neither an http body nor POST data is present, return error
+		$body = $this->request->getBody();
+		if(!$body && !$this->request->postVars()) {
+			$this->getResponse()->setStatusCode(204); // No Content
+			return 'No Content';
+		}
+		
+		if(!empty($body)) {
+			$data = $formatter->convertStringToArray($body);
+		} else {
+			// assume application/x-www-form-urlencoded which is automatically parsed by PHP
+			$data = $this->request->postVars();
+		}
+		
+		// @todo Disallow editing of certain keys in database
+		$data = array_diff_key($data, array('ID','Created'));
+		
+		$obj->update($data);
+		$obj->write();
+		
+		return $obj;
+	}
+	
 	protected function permissionFailure() {
 		// return a 401
 		$this->getResponse()->setStatusCode(403);
@@ -274,6 +358,33 @@ class RestfulServer extends Controller {
 		// return a 404
 		$this->getResponse()->setStatusCode(404);
 		return "That object wasn't found";
+	}
+	
+	protected function methodNotAllowed() {
+		$this->getResponse()->setStatusCode(405);
+		return "Method Not Allowed";
+	}
+	
+	protected function unsupportedMediaType() {
+		$this->response->setStatusCode(415); // Unsupported Media Type
+		return "Unsupported Media Type";
+	}
+	
+	protected function authenticate() {
+		if(!isset($_SERVER['PHP_AUTH_USER']) || !isset($_SERVER['PHP_AUTH_PW'])) return false;
+		
+		if($member = Member::currentMember()) return $member;
+		$member = MemberAuthenticator::authenticate(array(
+			'Email' => $_SERVER['PHP_AUTH_USER'], 
+			'Password' => $_SERVER['PHP_AUTH_PW'],
+		), null);
+		
+		if($member) {
+			$member->LogIn(false);
+			return $member;
+		} else {
+			return false;
+		}
 	}
 	
 }
